refactor: make suggested changes
