+++ .flake8
@@ -1,2 +1,3 @@
 [flake8]
-max-line-length = 89
+max-line-length = 88
+extend-ignore = E203
