+++ modules/ALFA/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from __future__ import absolute_import
 
 from .ALFA import MultiqcModule
